cni: allocate dynamic ip for pod veth peer
before this change only the host peer received a dynamiclly
allocated ip from the 10.0.10.0/24 subnet. the pod peer had
10.0.0.1 statically configured.

this is part to fix the following problem:

replies from the DNS server with 10.0.0.1 set as source would be
returned by the server, but never reach the host peer.

to address this is, we need to allocate a unique ip address
for each pod peer in order to create a routing table entry,
that routes packets matching the pod peers destination ip
to the corresponding host peer interface.