fixes required for passing acceptance criteria (#27)
this commit incorporates a bunch of changes required to successfully 
restore a checkpoint with functional networking over the platformd api.

this does not include refactorings, so there is some ugly code here and there.
cleanup will happen in seperate merge requests. 

## notable changes

### removal of static pod-side ip

before this change only the host peer received a dynamiclly allocated ip from 
the 10.0.10.0/24 subnet. the pod peer had10.0.0.1 statically configured.

this is part of the following problem:

replies from the DNS server with 10.0.0.1 set as source would be
returned by the server, but never reach the host peer.

to address this is, we need to allocate a unique ip address for each pod peer in 
order to create a routing table entry, that routes packets matching the pod peers 
destination ip to the corresponding host peer interface.

### envoy node name is statically configured

the previous approach straight up didn't work and since we will run only 1 envoy 
per host ever, this is no problem.