platformd: port allocation and workload status endpoint
allocate ports from a specified range and also provide status
endpoint to retrieve port and workload state. workload state
is currently not set.

for this to work we also had to implement an in-memory store to
keep track of all created workloads.

instead of passing `RunOptions`, we now pass the complete `Workload`
struct. this is because we now have to save the object before calling
`RunWorkload` to make it retrievable via `WorkloadStatus`, so the CNI
can obtain the allocated port.