allocate ports from a specified range to workloads (#34)
this commit allocates ports from a specified range and also provides a status
endpoint to retrieve port and workload state. workload state is currently not set.

this PR also introduces an ebpf map holding network configuration of a pod. 
this map is currently only used in the snat program, but will later also be used 
in the dnat program. this way we have a single map which holds all relevant
data about a pod.