db: return full chunk object with all flavors when creating instance
previously the flavors of a chunk were not returned. but since we
want to return the full chunk object when creating an instance,
this has to be done.

we are also inserting created_at and updated_at fields when creating
the objects, so values defined in our test fixtures are stored in the
database. this makes comparing the objects easier.