api: replace discovery api with instance service
returning workload objects from the discovery api does not make sense,
because converting an instance to a workload cannot be done 1 to 1.
details like labels, what network namespace mode should be used are
also properties the control plane should not be aware of. this is
part of the platformd domain.

this is why we remove the discovery api and provide a seperate service
in the chunk domain, which handles all things related to instances.

platformd will now query the `DiscoverInstances` endpoint and convert
the instances to workload objects.