add pod gc
this commit adds a gc loop that looks for pods that exceed
the specified attempt threshold and removes them.

this mechanism takes effect when the maximum workload creation
attempts are reached and the workload cannot be removed.
combinding all of this leads to zombie pods, because they are
unkown to both the control plane and platformd.