pod garbage collector (#47)
this commit adds a gc loop that looks for pods that exceed
the specified attempt threshold and removes them.

this mechanism takes effect when the maximum workload creation
attempts are reached and the workload cannot be removed.
combining all of this leads to zombie pods, because they are
unknown to both the control plane and platformd.