platformd: add checkpoint api (#112)
this adds the checkpoint api to platformd, which enables automated creation of 
checkpoints given a base image.

also the way how the gc works has been reworked. before there was pod_gc 
which only took care of deleting leftover pods only. now we also have the 
checkpointing, service which will also produce garbage. not only in the form 
of pods, but also in tarballs living in the filesystem.

other minor changes are:
- toolchain script now installs all components needed
  to run platformd.
- increase VM memory and disk size to support having more
  memory heavy containers running