controlplane: compute flavor diff by comparing path first instead of file content (#120)
this fixes a bug where a new file would be considered already
present if the content would be the same of any previously uploaded
file and there for not included in the new versions file list.

we fix this by taking the path into consideration first and then
checking the files hash.

other minor things to note:
- moved file package into shared internal folder as it will now also be used by the cli
- when retrieving a chunk, the latest flavor version will be the first in the versions list