ci: make functional tests run in parallel (#134)
using move functional tests into their own package based on component 
instead of using`t.Parallel()`, because this causes connection reset errors
on the testcontainers side. this could be investigated at some point, but
using the per-package approach is implemented faster for now. 

we save about 1 minute doing this, which is not much, but at least something :')