controlplane: adjust image worker to use s3 backend (#144)
the image worker now uses the s3 blob store instead of postgres. 
this is commit is a bit of a mess and contains a bunch of changes
that probably should have been their own pull requests.

anyway, we're now a step closer to finishing the file upload refactoring.
cli adjustements and checking whether the changeset upload has finished
are still needed.